requirements.txt
sqlalchemy~=2.0.31
psycopg2-binary
alembic~=1.13.2
fastapi~=0.111.1
pydantic~=2.8.2
python-dotenv~=1.0.1